+++ src/main/java/hype/interfaces/HRotatable.java
@@ -1,6 +1,13 @@
 package hype.interfaces;
 
 public interface HRotatable {
-	public float rotationRad();
-	public HRotatable rotationRad(float rad);
+	float rotationRad();
+	float rotationXRad();
+	float rotationYRad();
+	float rotationZRad();
+
+	HRotatable rotationRad(float rad); //retained for backwards compatibility
+	HRotatable rotationXRad(float radx);
+	HRotatable rotationYRad(float rady);
+	HRotatable rotationZRad(float radz);
 }